+++ api/test/trace/propagation/http_text_format_test.cc
@@ -1,33 +1,13 @@
-#include "opentelemetry/context/context.h"
-#include "opentelemetry/nostd/shared_ptr.h"
-#include "opentelemetry/nostd/span.h"
-#include "opentelemetry/nostd/string_view.h"
-#include "opentelemetry/trace/default_span.h"
-#include "opentelemetry/trace/noop.h"
-#include "opentelemetry/trace/span.h"
-#include "opentelemetry/trace/span_context.h"
-#include "opentelemetry/trace/trace_id.h"
-#include "opentelemetry/trace/tracer.h"
+#include "opentelemetry/context/runtime_context.h"
+#include "opentelemetry/trace/propagation/http_trace_context.h"
+#include "opentelemetry/trace/scope.h"
+#include "util.h"
 
 #include <map>
-#include <memory>
-#include <string>
 
 #include <gtest/gtest.h>
 
-#include "opentelemetry/trace/default_span.h"
-#include "opentelemetry/trace/propagation/http_text_format.h"
-#include "opentelemetry/trace/propagation/http_trace_context.h"
-
 using namespace opentelemetry;
-
-template <typename T>
-static std::string Hex(const T &id_item)
-{
-  char buf[T::kSize * 2];
-  id_item.ToLowerBase16(buf);
-  return std::string(buf, sizeof(buf));
-}
 
 static nostd::string_view Getter(const std::map<std::string, std::string> &carrier,
                                  nostd::string_view trace_type = "traceparent")
@@ -52,25 +32,12 @@ using MapHttpTraceContext =
 
 static MapHttpTraceContext format = MapHttpTraceContext();
 
-TEST(HTTPTextFormatTest, TraceIdBufferGeneration)
+TEST(TextMapPropagatorTest, TraceFlagsBufferGeneration)
 {
-  constexpr uint8_t buf[] = {1, 2, 3, 4, 5, 6, 7, 8, 8, 7, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff};
-  EXPECT_EQ(MapHttpTraceContext::GenerateTraceIdFromString("01020304050607080807aabbccddeeff"),
-            trace::TraceId(buf));
+  EXPECT_EQ(MapHttpTraceContext::TraceFlagsFromHex("00"), trace::TraceFlags());
 }
 
-TEST(HTTPTextFormatTest, SpanIdBufferGeneration)
-{
-  constexpr uint8_t buf[] = {1, 2, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff};
-  EXPECT_EQ(MapHttpTraceContext::GenerateSpanIdFromString("0102aabbccddeeff"), trace::SpanId(buf));
-}
-
-TEST(HTTPTextFormatTest, TraceFlagsBufferGeneration)
-{
-  EXPECT_EQ(MapHttpTraceContext::GenerateTraceFlagsFromString("00"), trace::TraceFlags());
-}
-
-TEST(HTTPTextFormatTest, NoSendEmptyTraceState)
+TEST(TextMapPropagatorTest, NoSendEmptyTraceState)
 {
   // If the trace state is empty, do not set the header.
   const std::map<std::string, std::string> carrier = {
@@ -85,7 +52,24 @@ TEST(HTTPTextFormatTest, NoSendEmptyTraceState)
   EXPECT_FALSE(carrier.count("tracestate") > 0);
 }
 
-TEST(HTTPTextFormatTest, PropagateInvalidContext)
+TEST(TextMapPropagatorTest, PropogateTraceState)
+{
+  const std::map<std::string, std::string> carrier = {
+      {"traceparent", "00-4bf92f3577b34da6a3ce929d0e0e4736-0102030405060708-01"},
+      {"tracestate", "congo=t61rcWkgMzE"}};
+  context::Context ctx1 = context::Context{
+      "current-span",
+      nostd::shared_ptr<trace::Span>(new trace::DefaultSpan(trace::SpanContext::GetInvalid()))};
+  context::Context ctx2                 = format.Extract(Getter, carrier, ctx1);
+  std::map<std::string, std::string> c2 = {};
+  format.Inject(Setter, c2, ctx2);
+
+  EXPECT_TRUE(carrier.count("traceparent") > 0);
+  EXPECT_TRUE(carrier.count("tracestate") > 0);
+  EXPECT_EQ(c2["tracestate"], "congo=t61rcWkgMzE");
+}
+
+TEST(TextMapPropagatorTest, PropagateInvalidContext)
 {
   // Do not propagate invalid trace context.
   std::map<std::string, std::string> carrier = {};
@@ -94,9 +78,10 @@ TEST(HTTPTextFormatTest, PropagateInvalidContext)
       nostd::shared_ptr<trace::Span>(new trace::DefaultSpan(trace::SpanContext::GetInvalid()))};
   format.Inject(Setter, carrier, ctx);
   EXPECT_TRUE(carrier.count("traceparent") == 0);
+  EXPECT_TRUE(carrier.count("tracestate") == 0);
 }
 
-TEST(HTTPTextFormatTest, SetRemoteSpan)
+TEST(TextMapPropagatorTest, SetRemoteSpan)
 {
   const std::map<std::string, std::string> carrier = {
       {"traceparent", "00-4bf92f3577b34da6a3ce929d0e0e4736-0102030405060708-01"}};
@@ -114,12 +99,14 @@ TEST(HTTPTextFormatTest, SetRemoteSpan)
   EXPECT_EQ(span->GetContext().IsRemote(), true);
 }
 
-TEST(HTTPTextFormatTest, GetCurrentSpan)
+TEST(TextMapPropagatorTest, GetCurrentSpan)
 {
   constexpr uint8_t buf_span[]  = {1, 2, 3, 4, 5, 6, 7, 8};
   constexpr uint8_t buf_trace[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16};
+
+  auto trace_state = trace::TraceState::FromHeader("congo=t61rcWkgMzE");
   trace::SpanContext span_context{trace::TraceId{buf_trace}, trace::SpanId{buf_span},
-                                  trace::TraceFlags{true}, false};
+                                  trace::TraceFlags{true}, false, trace_state};
   nostd::shared_ptr<trace::Span> sp{new trace::DefaultSpan{span_context}};
 
   // Set `sp` as the currently active span, which must be used by `Inject`.
@@ -128,4 +115,33 @@ TEST(HTTPTextFormatTest, GetCurrentSpan)
   std::map<std::string, std::string> headers = {};
   format.Inject(Setter, headers, context::RuntimeContext::GetCurrent());
   EXPECT_EQ(headers["traceparent"], "00-0102030405060708090a0b0c0d0e0f10-0102030405060708-01");
+  EXPECT_EQ(headers["tracestate"], "congo=t61rcWkgMzE");
+}
+
+TEST(TextMapPropagatorTest, InvalidIdentitiesAreNotExtracted)
+{
+  std::vector<std::string> traces = {
+      "ff-0af7651916cd43dd8448eb211c80319c-b9c7c989f97918e1-01",
+      "00-0af7651916cd43dd8448eb211c80319c1-b9c7c989f97918e1-01",
+      "00-0af7651916cd43dd8448eb211c80319c-b9c7c989f97918e11-01",
+      "0-0af7651916cd43dd8448eb211c80319c-b9c7c989f97918e1-01",
+      "00-0af7651916cd43dd8448eb211c80319c-b9c7c989f97918e1-0",
+      "00-0af7651916cd43dd8448eb211c8031-b9c7c989f97918e1-01",
+      "00-0af7651916cd43dd8448eb211c80319c-b9c7c989f97-01",
+      "00-1-1-00",
+      "00--b9c7c989f97918e1-01",
+      "00-0af7651916cd43dd8448eb211c80319c1--01",
+      "",
+      "---",
+  };
+
+  for (auto &trace : traces)
+  {
+    const std::map<std::string, std::string> carrier = {{"traceparent", trace}};
+    context::Context ctx1                            = context::Context{};
+    context::Context ctx2                            = format.Extract(Getter, carrier, ctx1);
+
+    auto span = trace::propagation::detail::GetCurrentSpan(ctx2);
+    EXPECT_FALSE(span.IsValid());
+  }
 }